cvat/requirements/base.txt
click==7.1.2
Django==3.1.1
django-appconf==1.0.4
django-auth-ldap==2.2.0
django-cacheops==5.0.1
django-compressor==2.4
django-rq==2.3.2
EasyProcess==0.3
Pillow==7.2.0
numpy==1.18.5
python-ldap==3.3.1
pytz==2020.1
pyunpack==0.2.1
rcssmin==1.0.6
redis==3.5.3
rjsmin==1.1.0
requests==2.24.0
rq==1.5.1
rq-scheduler==0.10.0
sqlparse==0.3.1
django-sendfile==0.3.11
dj-pagination==2.5.0
python-logstash==0.4.6
django-revproxy==0.10.0
rules==2.2
GitPython==3.1.8
coreapi==2.3.3
django-filter==2.3.0
Markdown==3.2.2
djangorestframework==3.11.1
Pygments==2.6.1
drf-yasg==1.17.1
Shapely==1.7.1
pdf2image==1.14.0
django-rest-auth[with_social]==0.9.5
cython==0.29.21
opencv-python==4.4.0.42
h5py==2.10.0
django-cors-headers==3.5.0
furl==2.1.0
av==8.0.2 --no-binary=av
tensorflow==2.2.1 # Optional requirement of Datumaro
# The package is used by pyunpack as a command line tool to support multiple
# archives. Don't use as a python module because it has GPL license.
patool==1.12
diskcache==5.0.2
git+https://github.com/openvinotoolkit/datumaro@v0.1.3